remove unnecessary test suite file
